Ambient music ticker AoD: make it compatibile with double tap sensor
also don't wake up on skip track events when pulsing Ambient

TLDR: some devices like the Pixel2xl have a specific double tap sensor for AoD
that handles the touches in this screen state and nothing is dispatched
to StatusBarWindowView, and it's enabled by the double tap to wake
option available in Settings - System - Gestures.

We can use the related onDoubleTap callback that Google already added
on purpose to manage double taps  on the AmbientContainer (Now Playing on
Android P) when AoD is showing, to trigger the skip track event we need.

Signed-off-by: Aston-Martinn <advaithbhat9@gmail.com>