Smart Pixels: Bypass check for obscured window
With the commit below Google added a check to not allow for a permission
to be granted if the window is obscured aka a screen filter is in play.
This check made sense because it protects the user from the user if said
user installs some shady app.

https://github.com/DirtyUnicorns/android_frameworks_base/commit/830d54f2c75bee8bda4c3bdd3d62c9f2f58934b7

The problem here is that it also disallows SmartPixels (a system mod). So this
commit checks for whether SmartPixels is enabled and allow the user to interact
with the window. If SmartPixels is not enabled then is business as usual and Google's
check will continue to be in play.

Test:
- Connect to a computer
- Enable SmartPixels
- Enable Developer options
- Open Settings/Developer options
- Revoke USB debugging authorizations
- Flip the USB debugging switch on
- Attempt to grant your computer USB debugging

Change-Id: I08f601e4f772b9d56fbacc54138cff49b03c465b
Signed-off-by: Adin Kwok <adin.kwok@carbonrom.org>
Signed-off-by: Aston-Martinn <advaithbhat9@gmail.com>