Keyguard: ReInflate views when density or font scale change
Previously the keyguard's statusbar was not fully adapting to the density or
font scale changes. Either the battery percent wouldn't stick or another icon
would not change. This was fixed by telling the user to just restart systemui
but that wasn't a real solution. This ensures that the statusbar view is indeed
adapting accordingly.

Test:
- Go into Settings/Display
- Change Font size or display size
- Turn your screen off and then on
- Look at the statusbar system icons

Change-Id: I435fdad1dada3bb0674e9fd5ee530e71e0516af0
Signed-off-by: Aston-Martinn <advaithbhat9@gmail.com>